__attribute__((constructor)) solution which works only with clang but not gcc

+++ checked_division/include/checked_division/checked_division.h
@@ -1,6 +1,6 @@
 #ifndef CHECKED_DIVISION_H
 #define CHECKED_DIVISION_H
 
-double divide(double x, double y);
+int divide(int x, int y);
 
 #endif
